+++ src/main/java/org/eclipse/tractusx/valueaddedservice/service/logic/ShareLogicService.java
@@ -45,13 +45,16 @@ public class  ShareLogicService {
     @Autowired
     ExternalBusinessPartnersLogicService externalBusinessPartnersLogicService;
 
+    @Autowired
+    BusinessPartnersLogicService businessPartnersLogicService;
+
     public List<ShareDTO> findRatingsScoresForEachBpn(List<DataSourceDTO> datasource, List<BusinessPartnerDTO> businessPartnerToMap, CompanyUserDTO companyUser,
                                                       String token,List<String> roles) {
 
         List<ShareDTO> shareDTOSList = new ArrayList<>();
         
         List<BusinessPartnerDTO> companyBusinessPartnerDTOS;
-        companyBusinessPartnerDTOS = externalBusinessPartnersLogicService.getExternalBusinessPartners(companyUser,token,roles);
+        companyBusinessPartnerDTOS = businessPartnersLogicService.getExternalBusinessPartners(companyUser,token,roles);
 
         businessPartnerToMap.forEach(bp -> {
             if(bp.getCountry() == null){